pcb layout: move some things around
This mostly involved swapping the transistor with the output coupling
capacitor so that the amplifier and feedback was on the left, and the
output stages were on the right.  I also rotated one resistor so the
nets were closer to each other, and swapped a few of the traces to other
layers of the PCB so that they could be shorter as well.

Overall, the result is pretty much just shorter traces and beefier
power-rail-fill in spots.